testing custom check function
